[chore] #151 - 팟 상세 업로더 정보를 셀 내부로 통합
